# server/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  jeri_blog_internal_dto.CreateTagRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        type: string
    required:
    - name
    type: object
...
  jeri_blog_internal_dto.MenuCreateRequest:
    properties:
      icon:
        description: 图标
        maxLength: 500
        type: string
      is_enabled:
        description: 是否启用
        type: boolean
      parent_id:
        description: 父菜单ID，NULL表示主菜单
        type: integer
      sort:
        description: 排序，范围1-10
        maximum: 10
        minimum: 1
        type: integer
      title:
        description: 菜单标题
        maxLength: 100
        minLength: 1
        type: string
      type:
        description: 菜单类型
        enum:
        - aggregate
        - navigation
        - footer
        type: string
      url:
        description: 链接地址（主菜单可为空）
        maxLength: 500
        type: string
    required:
    - title
    - type
    type: object
  jeri_blog_internal_dto.MenuDeleteRequest:
    properties:
      children_action:
        description: 子菜单处理方式：delete-删除子菜单，upgrade-升级为主菜单
        enum:
        - delete
        - upgrade
        type: string
    type: object
  jeri_blog_internal_dto.MenuResponse:
    properties:
      icon:
        type: string
      id:
        type: integer
      is_enabled:
        type: boolean
      parent_id:
        type: integer
      sort:
        type: integer
      title:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  jeri_blog_internal_dto.MenuTreeNode:
    properties:
      children:
        description: 子菜单列表
        items:
          $ref: '#/definitions/jeri_blog_internal_dto.MenuTreeNode'
        type: array
      icon:
        type: string
      id:
        type: integer
      is_enabled:
        type: boolean
      parent_id:
        type: integer
      sort:
        type: integer
      title:
        type: string
      type:
        type: string
      url:
        type: string
    type: object
  jeri_blog_internal_dto.MenuUpdateRequest:
    properties:
      icon:
        description: 图标
        maxLength: 500
        type: string
      is_enabled:
        description: 是否启用
        type: boolean
      parent_id:
        description: 父菜单ID
        type: integer
      sort:
        description: 排序
        maximum: 10
        minimum: 1
        type: integer
      title:
        description: 菜单标题
        maxLength: 100
        minLength: 1
        type: string
      type:
        description: 菜单类型
        enum:
        - aggregate
        - navigation
        - footer
        type: string
      url:
        description: 链接地址
        maxLength: 500
        type: string
    required:
    - title
    - type
    type: object
...
  jeri_blog_internal_dto.TagForWebResponse:
    properties:
      count:
        type: integer
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      url:
        type: string
    type: object
...
  jeri_blog_internal_dto.UpdateTagRequest:
    properties:
      description:
        maxLength: 200
        type: string
      name:
        maxLength: 50
        type: string
    required:
    - name
    type: object
...
  jeri_blog_internal_model.Tag:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      slug:
        type: string
      updated_at:
        type: string
    type: object
...
  jeri_blog_pkg_response.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
info:
  contact:
    email: 416685476@qq.com
    name: Jerion
  description: 一个基于 Go 语言的现代化博客后端服务
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Jeri-Server
  version: v1
paths:
...
  /admin/menus:
    get:
      consumes:
      - application/json
      parameters:
      - description: '菜单类型: aggregate/navigation/footer'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/jeri_blog_internal_dto.MenuTreeNode'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单树
      tags:
      - 菜单管理
    post:
      consumes:
      - application/json
      parameters:
      - description: 菜单信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.MenuCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.MenuResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 创建菜单
      tags:
      - 菜单管理
  /admin/menus/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 删除选项
        in: body
        name: request
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.MenuDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除菜单
      tags:
      - 菜单管理
    get:
      consumes:
      - application/json
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.MenuResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 获取菜单详情
      tags:
      - 菜单管理
    put:
      consumes:
      - application/json
      parameters:
      - description: 菜单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 菜单信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.MenuUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.MenuResponse'
              type: object
      security:
      - BearerAuth: []
      summary: 更新菜单
      tags:
      - 菜单管理
...
  /admin/tags:
    get:
      consumes:
      - application/json
      description: 获取所有标签用于管理
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量（不传则返回全部）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 标签列表（管理）
      tags:
      - 标签管理
    post:
      consumes:
      - application/json
      description: 创建新标签，自动生成 slug
      parameters:
      - description: 标签信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.CreateTagRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_model.Tag'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 创建标签
      tags:
      - 标签管理
  /admin/tags/{id}:
    delete:
      consumes:
      - application/json
      description: 软删除标签
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 删除标签
      tags:
      - 标签管理
    get:
      consumes:
      - application/json
      description: 通过 ID 获取标签信息
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 标签详情（管理）
      tags:
      - 标签管理
    put:
      consumes:
      - application/json
      description: 修改标签信息
      parameters:
      - description: 标签 ID
        in: path
        name: id
        required: true
        type: integer
      - description: 标签信息
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/jeri_blog_internal_dto.UpdateTagRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_model.Tag'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      security:
      - BearerAuth: []
      summary: 更新标签
      tags:
      - 标签管理
...
  /menus:
    get:
      consumes:
      - application/json
      parameters:
      - description: '菜单类型: aggregate/navigation/footer (不传则返回所有类型)'
        in: query
        name: type
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/jeri_blog_internal_dto.MenuTreeNode'
                  type: array
              type: object
      summary: 获取菜单树
      tags:
      - 菜单
...
  /tags:
    get:
      consumes:
      - application/json
      description: 获取所有标签
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      - description: 每页数量（不传则返回全部）
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      summary: 标签列表
      tags:
      - 标签
  /tags/{slug}:
    get:
      consumes:
      - application/json
      description: 通过 slug 获取标签信息
      parameters:
      - description: 标签 slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/jeri_blog_pkg_response.Response'
            - properties:
                data:
                  $ref: '#/definitions/jeri_blog_internal_dto.TagForWebResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/jeri_blog_pkg_response.Response'
      summary: 标签详情
      tags:
      - 标签
...
securityDefinitions:
  BearerAuth:
    description: 在请求头中添加 Bearer Token，格式：Bearer {token}
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"